Refactor: Move data classes to separate file
still broken
Moves `ArtifactColor`, `ArtifactSequence`, `Coords`, `DateMs`, and `MotorPositions` classes from `AutoDataStorageUtils.java` to a new file `DecodeDataTypes.java` in the `utils` package.

This refactoring improves code organization by separating data structures from utility functions.

Additionally, the `exampleAutoData.json` file is moved to the `utils` package to maintain consistency.
